backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
starlette
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.1.3
numpy==1.26.2
tqdm==4.66.1
transformers==4.38.2
pydantic-settings==2.1.0
scikit-learn==1.3.2
python-multipart==0.0.6
pydantic==2.5.2
fastembed==0.6.1
fastapi==0.104.1
uvicorn==0.24.0
starlette
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.1.3
numpy==1.26.2
tqdm==4.66.1
transformers==4.38.2
pydantic-settings==2.1.0
scikit-learn==1.3.2
python-multipart==0.0.6
pydantic==2.5.2
fastembed==0.6.1